Ref #19: Database callback fixes and minor improvements
I had to add the user field to get it to connect. Though,
I can't remember if it was alread there all of a sudden? Yick,
who knows.

Had to make the getUserInfo function user a proper replacement token.
Implemented the get user by email for verification.
The insert user needed a password field and to take proper flags.
Enjoyers needed password field for verification.